made code slightly more pytho2 friendly
