+++ test/completion/pep0484.py
@@ -157,70 +157,3 @@ Y = int
 def just_because_we_can(x: "flo" + "at"):
     #? float()
     x
-
-import typing
-def we_can_has_sequence(
-        p: typing.Sequence[int],
-        q: typing.Sequence[B],
-        r: "typing.Sequence[int]",
-        s: typing.Sequence["int"],
-        t: typing.MutableSequence[dict],
-        u: typing.List[float]):
-    #? ["count"]
-    p.c
-    #? int()
-    p[1]
-    #? ["count"]
-    q.c
-    #? B()
-    q[1]
-    #? ["count"]
-    r.c
-    #? int()
-    r[1]
-    #? ["count"]
-    s.c
-    #? int()
-    s[1]
-    #? []
-    s.a
-    #? ["append"]
-    t.a
-    #? dict()
-    t[1]
-    #? ["append"]
-    u.a
-    #? float()
-    u[1]
-
-def iterators(
-        ps: typing.Iterable[int],
-        qs: typing.Iterator[str],
-        rs: typing.Sequence["B"],
-        ts: typing.AbstractSet["float"]):
-    for p in ps:
-        #? int()
-        p
-    #?
-    next(ps)
-    for q in qs:
-        #? str()
-        q
-    #? str()
-    next(qs)
-    for r in rs:
-        #? B()
-        r
-    #?
-    next(rs)
-    for t in ts:
-        #? float()
-        t
-
-def sets(
-        p: typing.AbstractSet[int],
-        q: typing.MutableSet[float]):
-    #? []
-    p.a
-    #? ["add"]
-    q.a
